requirements.txt
discord.py
python-dotenv
solana
web3
openai
tweepy